+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `Stream::consume_reference` will return only `InvalidReference` error from now.
 - `Error::InvalidTokenWithCause` merged into `Error::InvalidToken`.
 - `Stream::gen_error_pos_from` does not require `mut self` from now.
+- `StreamError::InvalidChar` is a struct now.
 
 ### Removed
 - `failure` dependency.